requirements.txt
requests>=2.31
psycopg2-binary>=2.9
pydantic>=2.0
pytest>=8.0